Split off text-editing into separate module
