+++ parley/src/inputs/style/styleset.rs
@@ -12,9 +12,9 @@ type StyleProperty<Brush> = crate::inputs::StyleProperty<'static, Brush>;
 /// A long-lived collection of [`StyleProperties`](super::StyleProperty), containing at
 /// most one of each property.
 ///
-/// This is used by [`PlainEditor`](crate::outputs::editor::PlainEditor) to provide a reasonably ergonomic
+/// This is used by [`PlainEditor`](crate::editing::PlainEditor) to provide a reasonably ergonomic
 /// mutable API for styles applied to all text managed by it.
-/// This can be accessed using [`PlainEditor::edit_styles`](crate::outputs::editor::PlainEditor::edit_styles).
+/// This can be accessed using [`PlainEditor::edit_styles`](crate::editing::PlainEditor::edit_styles).
 ///
 /// These styles do not have a corresponding range, and are generally unsuited for rich text.
 #[derive(Clone, Debug)]
