+++ parley/src/tests/utils/cursor_test.rs
@@ -1,8 +1,9 @@
 // Copyright 2024 the Parley Authors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
+use crate::editing::Cursor;
 use crate::inputs::{FontContext, LayoutContext};
-use crate::outputs::{Affinity, Cursor, Layout};
+use crate::outputs::{Affinity, Layout};
 
 // Note: This module is only compiled when running tests, which requires std,
 // so we don't have to worry about being no_std-compatible.
